# api-docs/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "Guardian"
  description: "The Guardian is a modular open-source solution that includes best-in-class identity management and decentralized ledger technology (DLT) libraries. At the heart of the Guardian solution is a sophisticated Policy Workflow Engine (PWE) that enables applications to offer a requirements-based tokenization implementation."
  version: "2.9.3"
  contact:
    name: "API developer"
    url: "https://envisionblockchain.com"
    email: "info@envisionblockchain.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
- url: /api/v1
  description: version 1.0
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
...
    Token:
      type: object
      required:
        - changeSupply
        - decimals
        - enableAdmin
        - enableKYC
        - enableFreeze
        - enableWipe
        - initialSupply
        - tokenName
        - tokenSymbol
        - tokenType
      properties:
        changeSupply:
          type: boolean
        decimals:
          type: string
        enableAdmin:
          type: boolean
        enableFreeze:
          type: boolean
        enableKYC:
          type: boolean
        enableWipe:
          type: boolean
        initialSupply:
          type: string
        tokenName:
          type: string
        tokenSymbol:
          type: string
        tokenType:
          type: string
    TokenInfo:
      type: object
      required:
        - id
        - tokenId
        - tokenName
        - tokenSymbol
        - tokenType
        - decimals
        - associated
        - balance
        - frozen
        - kyc
      properties:
        id:
          type: string
        tokenId:
          type: string
        tokenName:
          type: string
        tokenSymbol:
          type: string
        tokenType:
          type: string
        decimals:
          type: string
        associated:
          type: boolean
        balance:
          type: string
        frozen:
          type: boolean
        kyc:
          type: boolean
        enableAdmin:
          type: boolean
        enableKYC:
          type: boolean
        enableFreeze:
          type: boolean
        enableWipe:
          type: boolean
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: number
        message:
          type: string
...
    Task:
      type: object
      properties:
        taskId:
          type: string
        expectation:
          type: number
...
paths:
...
  /tokens:
    get:
      tags:
        - tokens
      description: Returns all tokens. For the Standard Registry role it returns only the list of tokens, for other users it also returns token balances as well as the KYC, Freeze, and Association statuses. Not allowed for the Auditor role.
      security:
      - bearerAuth: []
      summary: Return a list of tokens.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/TokenInfo'
                    - type: object
                      properties:
                        policies:
                          type: array
                          items:
                            type: string
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - tokens
      description: Creates a new token. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      summary: Creates a new token.
      requestBody:
        description: Object that contains token information.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Token"
      responses:
        201:
          description: Successful operation.
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/TokenInfo'
                    - type: object
                      properties:
                        policies:
                          type: array
                          items:
                            type: string
        400:
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/push:
    post:
      tags:
        - tokens
      description: Creates a new token. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Creates a new token.
      requestBody:
        description: Object that contains token information.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Token"
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{tokenId}/{username}/info:
    get:
      tags:
        - tokens
      description: Returns user information for the selected token. Only users with the Standard Registry role are allowed to make the request.
      security:
      - bearerAuth: []
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
        400:
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{tokenId}/associate:
    put:
      tags:
        - tokens
      description: Associates the user with the provided Hedera token. Only users with the Installer role are allowed to make the request.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
      security:
      - bearerAuth: []
      summary: Associates the user with the provided Hedera token.
      responses:
        200:
          description: Successful operation.
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/push/{tokenId}/associate:
    put:
      tags:
        - tokens
      description: Associates the user with the provided Hedera token. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Associates the user with the provided Hedera token.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{tokenId}/dissociate:
    put:
      tags:
        - tokens
      description: Disassociates the user with the provided Hedera token. Only users with the Installer role are allowed to make the request.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
      security:
      - bearerAuth: []
      summary: Associate the user with the provided Hedera token.
      responses:
        200:
          description: Successful operation.
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/push/{tokenId}/dissociate:
    put:
      tags:
        - tokens
      description: Disassociates the user with the provided Hedera token. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Disassociates the user with the provided Hedera token.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /tokens/{tokenId}/{username}/freeze:
    put:
      tags:
        - tokens
      description: Freezes transfers of the specified token for the user. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username.
      security:
      - bearerAuth: []
      summary: Freeze transfers of the specified token for the user.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
        400:
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/push/{tokenId}/{username}/freeze:
    put:
      tags:
        - tokens
      description: Freezes transfers of the specified token for the user. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Freeze transfers of the specified token for the user.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/{tokenId}/{username}/unfreeze:
    put:
      tags:
        - tokens
      description: Unfreezes transfers of the specified token for the user. Only users with the Standard Registry role are allowed to make the request.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username.
      security:
      - bearerAuth: []
      summary: Unfreezes transfers of the specified token for the user.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfo'
        400:
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /tokens/push/{tokenId}/{username}/unfreeze:
    put:
      tags:
        - tokens
      description: Unfreezes transfers of the specified token for the user. Only users with the Standard Registry role are allowed to make the request.
      security:
        - bearerAuth: []
      summary: Unfreezes transfers of the specified token for the user.
      parameters:
        - in: path
          name: tokenId
          schema:
            type: string
          required: true
          description: Token ID.
        - in: path
          name: username
          schema:
            type: string
          required: true
          description: Username.
      responses:
        200:
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        401:
          description: Unauthorized.
        403:
          description: Forbidden.
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...